\name{MAP_DP}
\docType{data}
\alias{MAP_DP}
\title{SecDb mapping from delivery point to parent.}
\description{
  In SecDb it's used during position reporting.  Used in PM
  package for node marking. 

  Run \code{PM:::.makeSecDbDelivPt.MAP} from \code{PM:::utils.R} to
  update the data.

  Has data for NEPOOL and NYPP.
}
\usage{data(MAP_DP)}
\format{One link per row.}
\source{SecDb}
\keyword{datasets}

