\name{PM-package}
\alias{PM-package}
\alias{PM}
\docType{package}
\title{
A collection of utilities for daily PM activities.
}
\details{
\tabular{ll}{
Package: \tab PM\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2009-01-30\cr
License: \tab CEG\cr
}
}
\author{
Adrian Dragulescu

Maintainer: Adrian Dragulescu <adrian.dragulescu@constellation.com>
}

\description{
  A collection of utility functions for daily PM activities
  (with a focus on NEPOOL).  
}
\keyword{ package }

\examples{

# I need to put some examples here ... 

# Several data sets are available.

# a SecDb map of delivery points to their parents:
data(MAP_DP)
head(MAP_DP)
#   mkt parentDelivPt       delivPt
# 1 PWX        ANCSVC        ANCSVC
# 2 PWX          ICAP          ICAP
# 3 PWX           PTF           PTF
# 4 PWX         SHAPE         SHAPE
# 5 PWX SELLERSCHOICE SELLERSCHOICE
# 6 PWX            SC            SC



}
